% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpace.R
\docType{package}
\name{BayesSpace}
\alias{BayesSpace-package}
\alias{BayesSpace}
\title{BayesSpace: A package for processing spatial transcriptomes}
\description{
Tools for clustering and enhancing the resolution of spatial gene expression
experiments. BayesSpace clusters a low-dimensional representation of the gene
expression matrix, incorporating a spatial prior to encourage neighboring
spots to cluster together. The method can enhance the resolution of the
low-dimensional representation into "sub-spots", for which features such as
gene expression or cell type composition can be imputed.
}
\details{
For an overview of the functionality provided by the package, please see the
vignette:
\code{vignette("BayesSpace", package="BayesSpace")}
}
\seealso{
Useful links:
\itemize{
  \item \url{edward130603.github.io/BayesSpace}
  \item Report bugs at \url{https://github.com/edward130603/BayesSpace/issues}
}

}
\author{
\strong{Maintainer}: Senbai Kang \email{senbai.kang@chuv.ch}

Authors:
\itemize{
  \item Edward Zhao \email{edward130603@gmail.com}
  \item Matt Stone \email{mstone@fredhutch.org}
}

Other contributors:
\itemize{
  \item Xing Ren \email{xren2@fredhutch.org} [contributor]
  \item Raphael Gottardo \email{rgottard@fredhutch.org} [contributor]
}

}
\keyword{internal}
