% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Spatial clustering}
\usage{
cluster(
  Y,
  positions,
  neighborhood.radius,
  q,
  model = c("normal", "t"),
  precision = c("equal", "variable"),
  z0 = rep(1, nrow(Y)),
  mu0 = colMeans(Y),
  lambda0 = diag(0.01, nrow = ncol(Y)),
  gamma = 2,
  alpha = 1,
  beta = 0.01,
  nrep = 1000
)
}
\arguments{
\item{Y}{A matrix or dataframe with 1 row per spot and 1 column per outcome 
(e.g. principal components).}

\item{positions}{A matrix or dataframe with two columns (x, y) that gives
the spatial coordinates of each spot.}

\item{neighborhood.radius}{The maximum (L1) distance for two spots to be
considered neighbors.}

\item{q}{The number of clusters.}

\item{model}{Error model. ("normal" or "t")}

\item{precision}{Covariance structure. ("equal" or "variable" for EEE and 
VVV covariance models, respectively.)}

\item{mu0}{Prior mean hyperparameter for mu.}

\item{lambda0}{Prior precision hyperparam for mu.}

\item{gamma}{Smoothing parameter. (Values in range of 1-3 seem to work well.)}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}

\item{nrep}{The maximum number of MCMC iterations.}

\item{init}{Initial cluster assignments (z's). Must be a vector of length
equal to the number of rows of Y and positions.}
}
\value{
List of parameter values (`z`, `mu`, `lambda`) and model 
        log-likelihoods (`plogLik`) at each MCMC iteration, along with final
        cluster labels (`labels`)
}
\description{
Backend calls iterate() which is written in Rcpp
}
\details{
TODO describe method in detail
}
