% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{cluster.results}
\alias{cluster.results}
\title{Get the cluster assignments of the BayesSpace}
\usage{
cluster.results(
  Y,
  q,
  spots.neighbors,
  init,
  mu0 = colMeans(Y),
  Lambda0 = diag(ncol(Y)) * 0.01,
  gamma = 3,
  alpha = 1,
  beta = 0.01,
  ncores = 2,
  n.rep = 1000,
  save_para = FALSE
)
}
\arguments{
\item{Y}{The PCs of SingleCellExperiment}

\item{q}{The numbers of clusters}

\item{spots.neighbors}{List of neighbor indices for each spot}

\item{init}{Vector of initial cluster assignments}

\item{mu0}{Prior mean hyperparameter for mu. If not provided, mu0 is set to
the mean of PCs over all spots.}

\item{Lambda0}{Prior precision hyperparam for mu. If not provided, lambda0
is set to a diagonal matrix \eqn{0.01 I}.}

\item{gamma}{Smoothing parameter. Defaults to 2 for \code{platform="ST"} and
3 for \code{platform="Visium"}.}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}

\item{ncores}{The number of cores used for parallel calculation.}

\item{n.rep}{The number of MCMC iterations}

\item{save_para}{If true, return the parameter mu, w and Lambda apart from z.}
}
\value{
If \code{save_para = FALSE}, return the vector of cluster assignments.
If \code{save_para = TRUE}, return the list of clustering parameter values apart from the vector of cluster assignments.
}
\description{
Get the cluster assignments of the BayesSpace
}
\keyword{internal}
