% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{Plot spatial cluster assignments.}
\usage{
ClusterPlot(
  sce,
  label = "spatial.cluster",
  palette = NULL,
  platform = NULL,
  is.enhanced = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment. If \code{label} is specified and is a string,
it must exist as a column in \code{colData(sce)}.}

\item{label}{Labels used to color each spot. May be the name of a column in
\code{colData(sce)}, or a vector of discrete values.}

\item{palette}{Optional vector of hex codes to use for discrete spot values.}

\item{platform}{Spatial transcriptomic platform.}

\item{is.enhanced}{Used to determine spot layout.}
}
\value{
Returns a ggplot object.
}
\description{
Plot spatial cluster assignments.
}
