% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_neighborhood_radius}
\alias{compute_neighborhood_radius}
\title{Compute the distance between two neighboring spots}
\usage{
compute_neighborhood_radius(sce)
}
\value{
double radius
}
\description{
Compute the distance between two neighboring spots
}
