% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{deconvolve}
\alias{deconvolve}
\title{Wrapper around C++ \code{iterate_deconv()} function}
\usage{
deconvolve(
  Y,
  positions,
  xdist,
  ydist,
  scalef,
  q,
  init,
  nrep = 1000,
  model = "normal",
  platform = c("Visium", "ST"),
  verbose = TRUE,
  jitter_scale = 5,
  jitter_prior = 0.01,
  adapt.before = 100,
  mu0 = colMeans(Y),
  gamma = 2,
  lambda0 = diag(0.01, nrow = ncol(Y)),
  alpha = 1,
  beta = 0.01,
  cores = 1
)
}
\value{
List of enhancement parameter values at each iteration
}
\description{
Wrapper around C++ \code{iterate_deconv()} function
}
\keyword{internal}
