% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.adjust_hex_centers}
\alias{.adjust_hex_centers}
\title{Adjust hex spot positions so hexagons are adjacent to each other in plot}
\usage{
.adjust_hex_centers(spot_positions)
}
\description{
Spots are regular hexagons with one unit of horizontal distance
between centers
}
\keyword{internal}
