% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bsData}
\alias{.bsData}
\title{Access BayesSpace metadata}
\usage{
.bsData(sce, name, default = NULL, warn = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{name}{Metadata name}
}
\value{
Requested metadata
}
\description{
Access BayesSpace metadata
}
\keyword{internal}
