% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_dimred}
\alias{.check_dimred}
\title{Check for reducedDim features}
\usage{
.check_dimred(sce, name)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{name}{Names of reduced dimensions}
}
\value{
A list of combined reducedDim features
}
\description{
Check for reducedDim features
}
\keyword{internal}
