% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcChain.R
\name{.clean_chain}
\alias{.clean_chain}
\title{Tidy C++ outputs before writing to disk.}
\usage{
.clean_chain(out, method = c("cluster", "enhance"), thin = 100)
}
\arguments{
\item{out}{List returned by \code{cluster()} or \code{deconvolve()}.}

\item{method}{Whether the output came from clustering or enhancement.
(Different params are included in each.)}

\item{thin}{Thinning rate. Some enhanced parameters are thinned within C++
loop, others (\code{mu} and \code{Ychange}) need to be thinned afterwards.}
}
\value{
List with standardized parameters
}
\description{
1) Convert each parameter to matrix (n_iterations x n_indices) 
2) Add appropriate colnames 
3) Thin evenly (for enhance)
}
\keyword{internal}
