% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.compute_interspot_distances}
\alias{.compute_interspot_distances}
\title{Estimate the distance between two neighboring spots}
\usage{
.compute_interspot_distances(sce)
}
\arguments{
\item{sce}{SingleCellExperiment (must include array_row, array_col, pxl_row_in_fullres, pxl_col_in_fullres
in colData)}

\item{scale.factor}{Scale estimated L1 difference up by this amount.}
}
\value{
doubles xdist, ydist
}
\description{
Fit linear models between each image pixel coordinate and its corresponding
array coordinate to estimate the pixel distance between two spots along
each axis. Add these distances to estimate the L1 distance between two
spots, then add a small buffer.
}
\keyword{internal}
