% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVisium.R
\name{.extract_indices}
\alias{.extract_indices}
\title{Extract row and column indices of the count matrix from h5 file.}
\usage{
.extract_indices(idx, new.start, zero.based = TRUE)
}
\arguments{
\item{idx}{Row index of corresponding element in the non-zero count matrix.}

\item{new.start}{Index of the start of each column corresponding to
\code{idx} and the non-zero count matrix.}

\item{zero.based}{Whether the \code{} and \code{} are zero-based or not.
(By default is TRUE)}
}
\value{
List of row (i) and column (j) indices of the non-zero elements
    in the count matrix.
}
\description{
Extract row and column indices of the count matrix from h5 file.
}
\keyword{internal}
