% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_outline.R
\name{.find_connected_components}
\alias{.find_connected_components}
\title{Find connected components of each cluster in spatial context}
\usage{
.find_connected_components(sce, group = "spatial.cluster")
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{group}{Categorical column in \code{colData(sce)} to group spots by}
}
\value{
Vector of component memberships
}
\description{
Consider spots connected if they are neighbors and belong to same cluster
}
\keyword{internal}
