% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{.find_neighbors}
\alias{.find_neighbors}
\title{Find neighboring spots based on array coordinates}
\usage{
.find_neighbors(sce, platform, k)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{platform}{If "Visium", select six neighboring spots around center; if
"ST", select four adjacent spots.}

\item{k}{If "others", use k nearest neighbors}
}
\value{
\code{df_j} a list of neighbor indices (zero-indexed) for each spot
}
\description{
Find neighboring spots based on array coordinates
}
\keyword{internal}
