% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcChain.R
\name{.flatten_matrix_list}
\alias{.flatten_matrix_list}
\title{Convert a list of matrices to a single matrix, where each row is a flattened
matrix from the original list}
\usage{
.flatten_matrix_list(xs, ...)
}
\arguments{
\item{xs}{List of matrices}
}
\value{
Matrix
}
\description{
Convert a list of matrices to a single matrix, where each row is a flattened
matrix from the original list
}
\keyword{internal}
