% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.flip_axis}
\alias{.flip_axis}
\title{Whether to flip x and y axis to align the plot with the corresponding image.}
\usage{
.flip_axis(sce, platform)
}
\value{
A list indicates the multiplier for each axis.
}
\description{
Whether to flip x and y axis to align the plot with the corresponding image.
}
\keyword{internal}
