% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_outline.R
\name{.get_boundary_vertices}
\alias{.get_boundary_vertices}
\title{Select boundary vertices from a connected component}
\usage{
.get_boundary_vertices(vertices, max_neighbors = 2)
}
\arguments{
\item{max_neighbors}{Maximum number of neighboring spots for a spot to be
considered on the boundary.}

\item{component}{Subset of vertices from \link{\code{.make_vertices()}}}
}
\value{
A data.frame of x and y coordinates for each boundary vertex
}
\description{
A vertex appears once in each spot outline. Since adjacent spots coincide in
our plotting functions, we can determine whether the vertex is at the
boundary of the component by counting the number of spots it appears in. A
vertex that appears in the maximum possible number of spots at an
intersection (i.e. 3 for Visium and 4 for ST) is in the interior of the
component; any vertex that appears in fewer than the maximum is on the
boundary.
}
\keyword{internal}
