% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcChain.R
\name{.infer_param_dims}
\alias{.infer_param_dims}
\title{Infer original dimensions of parameter (per iteration) from colnames}
\usage{
.infer_param_dims(cnames)
}
\arguments{
\item{cnames}{List of column names}
}
\value{
Numeric vector (nrow, ncol)
}
\description{
Used to avoid writing colnames directly to HDF5 as attribute, which fails
for large parameters (e.g. Y)
}
\keyword{internal}
