% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{.init_cluster}
\alias{.init_cluster}
\title{Initialize cluster assignments}
\usage{
.init_cluster(
  Y,
  q,
  init = NULL,
  init.method = c("mclust", "mclust-hc", "kmeans", "teigen"),
  init.args = NULL,
  teigen.max.attempts = 10
)
}
\arguments{
\item{q}{Number of clusters}

\item{init}{Vector of initial cluster assignments}

\item{init.method}{Initialization clustering algorithm}

\item{teigen.max.attempts}{teigen may fail to converge depending on seed.
Sometimes it will converge if rerun. This controls the maximum number of
attempts before giving up.}

\item{sce}{SingleCellExperiment}

\item{inputs}{Results from \code{.prepare_inputs()}}
}
\value{
Vector of cluster assignments.
}
\description{
Initialize cluster assignments
}
\keyword{internal}
