% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_outline.R
\name{.is_left}
\alias{.is_left}
\title{Check if \code{pt} is to left of line determined by \code{x1} and \code{x2}}
\usage{
.is_left(x1, x2, pt)
}
\arguments{
\item{x1, x2, pt}{(x,y) coordinates}
}
\value{
\code{TRUE} if \code{pt} is to the left of \code{(x1, x2)} from the
  perspective of \code{x1}. Otherwise \code{FALSE}.
}
\description{
Used to determine if a vertex in the boundary is to the left or right of the
preceding path.
}
\details{
\code{pt} is to the left of \code{(x1, x2)} if the sign of the determinant of
\code{(x2 - x1)} and \code{(pt - x1)} (the lines drawn from \code{x1} to each
of its neighbors) is positive.
}
\keyword{internal}
