% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.list2vec}
\alias{.list2vec}
\title{Convert a list into vectors for easier output.}
\usage{
.list2vec(X, sep = "=", collapse = ",", use_names = TRUE)
}
\arguments{
\item{X}{A list.}
}
\value{
A vector converted from the input list \code{X}.
}
\description{
Convert a list into vectors for easier output.
}
\keyword{internal}
