% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.make_hex_spots}
\alias{.make_hex_spots}
\title{Make vertices for each hex spot}
\usage{
.make_hex_spots(cdata)
}
\value{
Table of (x.pos, y.pos, spot, fill); where \code{spot} groups the
  vertices outlining the spot's border
}
\description{
Make vertices for each hex spot
}
\keyword{internal}
