% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcChain.R
\name{.make_index_names}
\alias{.make_index_names}
\title{Make colnames for parameter indices.}
\usage{
.make_index_names(name, m = NULL, n = NULL, dim = 1)
}
\arguments{
\item{name}{Parameter name}

\item{m, n}{Dimensions of parameter (m=nrow, n=ncol)}

\item{dim}{Dimensionality of parameter (0=scalar, 1=vector, 2=matrix)}
}
\value{
List of names for parameter values
}
\description{
Scalar parameters are named \code{"name"}.
Vector parameters are named \code{"name[i]"}.
Matrix parameters are named \code{"name[i,j]"}.
}
\keyword{internal}
