% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlinePlot.R
\name{.make_outline}
\alias{.make_outline}
\title{Obtain vertices outlining a connected component}
\usage{
.make_outline(component)
}
\arguments{
\item{component}{Table with \code{(x.vertex, y.vertex)} coordinates}
}
\value{
Vertices from boundary of \code{component}, ordered for plotting with
  \code{geom_polygon()}
}
\description{
Obtain vertices outlining a connected component
}
\keyword{internal}
