% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.make_square_spots}
\alias{.make_square_spots}
\title{Make vertices for each square spot}
\usage{
.make_square_spots(cdata, fill.col = "spatial.cluster", scale.factor = 1)
}
\description{
Squares are simple, just mae a unit square at each array coordinate
}
\keyword{internal}
