% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{.make_subspot_coldata}
\alias{.make_subspot_coldata}
\title{Add subspot labels and offset row/col locations before making enhanced SCE.}
\usage{
.make_subspot_coldata(
  positions,
  sce,
  n_subspots_per,
  spot_neighbors,
  subspot_neighbors
)
}
\arguments{
\item{sce}{Original sce (to obtain number of spots and original row/col)}

\item{n_subspots_per}{Number of subspots per spot}

\item{cdata}{Table of colData (imagerow and imagecol; from deconv$positions)}
}
\value{
Data frame with added subspot names, parent spot indices, and offset
  row/column coordinates
}
\description{
Subspots are stored as (1.1, 2.1, 3.1, ..., 1.2, 2.2, 3.2, ...)
}
\keyword{internal}
