% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{.make_subspots}
\alias{.make_subspots}
\title{Define offsets and Manhattan distances for each subspot layout.}
\usage{
.make_subspots(
  platform,
  xdist,
  ydist,
  force = FALSE,
  nsubspots.per.edge = 3,
  tolerance = 1.05
)
}
\description{
Hex spots are divided into 6 triangular subspots, square spots are divided
into 9 squares. Offsets are relative to the spot center. A unit corresponds
to the diameter of a spot.
}
\details{
Manhattan distance is used here instead of Euclidean to avoid numerical
issues.
}
\keyword{internal}
