% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.make_vertices}
\alias{.make_vertices}
\title{Make vertices outlining spots/subspots for geom_polygon()}
\usage{
.make_vertices(sce, fill, platform, is.enhanced)
}
\arguments{
\item{sce}{SingleCellExperiment with row/col in colData}

\item{fill}{Name of a column in \code{colData(sce)} or a vector of values to
use as fill for each spot}

\item{platform}{"Visium" or "ST", used to determine spot layout}

\item{is.enhanced}{If true, \code{sce} contains enhanced subspot data instead
of spot-level expression. Used to determine spot layout.}
}
\value{
Table of (x.pos, y.pos, spot, fill); where \code{spot} groups the
  vertices outlining the spot's border
}
\description{
Make vertices outlining spots/subspots for geom_polygon()
}
\keyword{internal}
