% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{.merge_default_cluster_args}
\alias{.merge_default_cluster_args}
\title{When the user specifies additional arguments to a cluster initialization,
merge these with our default values, giving precedence to the user-specified
value.}
\usage{
.merge_default_cluster_args(defaults, init.args = NULL)
}
\arguments{
\item{defaults}{List of default function arguments}

\item{init.args}{List of user-specified function arguments}
}
\value{
List of arguments. Any arguments with default values are added to
  \code{init.args} if the user did not specify a value.
}
\description{
When the user specifies additional arguments to a cluster initialization,
merge these with our default values, giving precedence to the user-specified
value.
}
\keyword{internal}
