% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlinePlot.R
\name{.nudge_outline}
\alias{.nudge_outline}
\title{Nudge outline into interior of component, to avoid overlaps between clusters}
\usage{
.nudge_outline(boundary, nudge = 0.25)
}
\arguments{
\item{boundary}{Table of (x, y) coordinates}

\item{nudge}{Fraction of distance to midpoint of neighbors}
}
\value{
Table of nudged coordinates at \code{(x_nudge, y_nudge)}
}
\description{
Moves each boundary vertex \code{nudge} of the distance towards the midpoint
of its neighbors. Distance is negated if the midpoint is outside the boundary
}
