% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_inputs}
\alias{.prepare_inputs}
\title{Prepare cluster/deconvolve inputs from SingleCellExperiment object}
\usage{
.prepare_inputs(
  sce,
  subspots,
  use.dimred = list(PCA = seq_len(15)),
  use.subspot.dimred = NULL,
  jitter_prior = 0.3,
  calc.neighbors = TRUE,
  calc.init = TRUE,
  q = NULL,
  init = NULL,
  init.method = c("spatialCluster", "mclust", "kmeans"),
  positions = NULL,
  position.cols = c("pxl_col_in_fullres", "pxl_row_in_fullres"),
  xdist = NULL,
  ydist = NULL,
  platform = c("Visium", "ST"),
  verbose = FALSE
)
}
\value{
List of PCs, names of columns with x/y positions, and inter-spot
  distances
}
\description{
Prepare cluster/deconvolve inputs from SingleCellExperiment object
}
\keyword{internal}
