% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_subspot_coldata}
\alias{.prepare_subspot_coldata}
\title{Add subspot labels and offset row/col locations before making enhanced SCE.}
\usage{
.prepare_subspot_coldata(positions, sce, subspots)
}
\arguments{
\item{sce}{Original sce (to obtain number of spots and original row/col)}

\item{cdata}{Table of colData (imagerow and imagecol; from deconv$positions)}

\item{n_subspots}{Number of subspots per spot}
}
\value{
Data frame with added subspot names, parent spot indices, and offset
  row/column coordinates
}
\description{
Subspots are stored as (1.1, 2.1, 3.1, ..., 1.2, 2.2, 3.2, ...)
}
\keyword{internal}
