% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcChain.R
\name{.read_chain}
\alias{.read_chain}
\title{Load saved chain from disk.}
\usage{
.read_chain(h5.fname, params = NULL, is.enhanced = FALSE)
}
\arguments{
\item{h5.fname}{Path to hdf5 file containing chain}

\item{params}{List of parameters to read from file (will read all by default)}
}
\value{
MCMC chain, represented as a \code{coda::mcmc} object
}
\description{
Load saved chain from disk.
}
\keyword{internal}
