% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVisium.R
\name{.read_spot_pos}
\alias{.read_spot_pos}
\title{Load spot positions.}
\usage{
.read_spot_pos(dirname)
}
\arguments{
\item{dirname}{Path to spaceranger outputs of spatial pipeline, i.e., "outs/spatial".
This directory must contain a file for the spot positions at
\code{tissue_positions_list.csv} (before Space Ranger V2.0) or
\code{tissue_positions.csv} (since Space Ranger V2.0).}
}
\value{
Data frame of spot positions.
}
\description{
Load spot positions.
}
\keyword{internal}
