% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlinePlot.R
\name{.roll}
\alias{.roll}
\title{Shift a vector forwards or backwards}
\usage{
.roll(x, n = 1L)
}
\arguments{
\item{x}{Vector to roll.}

\item{n}{Number of elements to shift vector by. If positive, elements are
moved forwards; if negative, backwards.}
}
\value{

}
\description{
Like \code{dplyr::lag()} or \code{dplyr::lead()}, but values from the
head/tail of the vector are rolled over to the end/start instead of being
replaced with a default value.
}
\examples{
.roll(1:5)
.roll(1:5, -1)

}
\keyword{internal}
