% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.select_spot_positions}
\alias{.select_spot_positions}
\title{Helper to extract x, y, fill ID}
\usage{
.select_spot_positions(
  cdata,
  x = "col",
  y = "row",
  fill.col = "spatial.cluster"
)
}
\description{
Helper to extract x, y, fill ID
}
\keyword{internal}
