% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.select_subspot_positions}
\alias{.select_subspot_positions}
\title{Helper to pull out subspot position columns
Probably redundant with select_spot_positions above, but we need subspot.idx}
\usage{
.select_subspot_positions(
  cdata,
  x = "spot.col",
  y = "spot.row",
  fill.col = "spatial.cluster"
)
}
\value{
Dataframe of (x.pos, y.pos, fill) for each spot
}
\description{
Helper to pull out subspot position columns
Probably redundant with select_spot_positions above, but we need subspot.idx
}
\keyword{internal}
