% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlinePlot.R
\name{.trace_path}
\alias{.trace_path}
\title{Trace a path clockwise through a set of vertices}
\usage{
.trace_path(X, max_dist = 0.6)
}
\arguments{
\item{X}{matrix of (x, y) coordinates for vertices}

\item{max_dist}{Maximum distance for two vertices to be considered neighbors.
Defaults to 0.6 for Visium; TODO write for ST}
}
\value{
A vector of indices into the rows of \code{X}, ordered such that the
  coordinates of \code{X} may be traced into a polygon.
}
\description{
Used to arrange boundary vertices for \code{geom_path()}
}
\keyword{internal}
