% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{FeaturePlot}
\alias{FeaturePlot}
\title{Plot spatial gene expression}
\usage{
FeaturePlot(
  sce,
  feature,
  platform = NULL,
  is.enhanced = NULL,
  diverging = FALSE,
  low = NULL,
  high = NULL,
  mid = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment. If \code{label} is specified and is a string,
it must exist as a column in \code{colData(sce)}.}

\item{feature}{Feature vector used to color each spot. May be the name of a
gene/row in an assay of \code{sce}, or a vector of continuous values.}

\item{platform}{Spatial transcriptomic platform.}

\item{is.enhanced}{Used to determine spot layout.}

\item{diverging}{If true, use a diverging color gradient in
\code{featurePlot()} (e.g. when plotting a fold change) instead of a
sequential gradient (e.g. when plotting expression).}

\item{low, mid, high}{Optional hex codes for low, mid, and high values of the
color gradient used for continuous spot values.}
}
\value{
Returns a ggplot object.
}
\description{
Plot spatial gene expression
}
