% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{features}
\alias{features}
\title{Get predicted features}
\usage{
features(
  X.enhanced,
  X,
  Y,
  feature_names = rownames(Y),
  model = c("xgboost", "lm"),
  nrounds,
  train.n
)
}
\arguments{
\item{X.enhanced}{Enhanced PCs}

\item{X}{Original PCs}

\item{Y}{Original log expression.}

\item{feature_names}{List of genes/features to predict expression/values for.}

\item{model}{Model used to predict enhanced values.}

\item{nrounds}{Nonnegative integer to set the \code{nrounds} parameter for xgboost.
If \code{nrounds} is set to 0, the parameter will be tuned using a train-test split.
We recommend tuning \code{nrounds} for improved feature prediction, but note this will increase
runtime.}

\item{train.n}{Number of spots to use in the training dataset for tuning
nrounds. By default, 2/3 the total number of spots are used.}
}
\value{
Vector of enhanced features.
}
\description{
Get predicted features
}
