% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{find.neighbors}
\alias{find.neighbors}
\title{Find neighboring spots based on array coordinates}
\usage{
find.neighbors(sce, platform)
}
\arguments{
\item{sce}{SingleCellExpreriment}

\item{platform}{If "Visium", select six neighboring spots around center; if
"ST", select four adjacent spots.}
}
\value{
\code{spots.neighbors} a list of neighbor indices for each spot
}
\description{
Find neighboring spots based on array coordinates
}
\keyword{internal}
