% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_cluster.R
\name{find.neighbors.sub}
\alias{find.neighbors.sub}
\title{Compute pairwise distances between all subspots and return list of neighbors
for each subspot.}
\usage{
find.neighbors.sub(sub.positions, platform)
}
\arguments{
\item{sub.positions}{Data frame with added subplot names, parent spot indices, subplot indice,
and offset row/column coordinates.}

\item{platform}{if "Visium",  subplots with distances less than 0.7 are neighbors.}
}
\value{
\code{sub.neighbors} a list of neighbor indices for each subspot
}
\description{
Compute pairwise distances between all subspots and return list of neighbors
for each subspot.
}
\keyword{internal}
