% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Compute pairwise distances between all spots and return list of neighbors
for each spot.}
\usage{
find_neighbors(positions, radius, method = c("manhattan", "euclidean"))
}
\arguments{
\item{positions}{(n x 2) matrix of spot coordinates.}

\item{radius}{The maximum distance for two spots to be considered neighbors.}

\item{method}{Distance metric to use.}
}
\value{
List df_j, where df_j[[i]] is a vector of zero-indexed neighbors of i.
}
\description{
Compute pairwise distances between all spots and return list of neighbors
for each spot.
}
