% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hex_spots_vertices}
\alias{hex_spots_vertices}
\title{Make vertices for each hexagonal spot}
\usage{
hex_spots_vertices(sce, x = "col", y = "row", fill = "spatial.cluster")
}
\value{
Table of (x.pos, y.pos, spot, fill); where \code{spot} groups the
  vertices outlining the spot's border
}
\description{
Make vertices for each hexagonal spot
}
\keyword{internal}
