% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{imageFeaturePlot}
\alias{imageFeaturePlot}
\title{Plot features extracted from images.}
\usage{
imageFeaturePlot(
  sce,
  datatype = c("pca", "vae", "both"),
  res = c("spot", "subspot", "both"),
  d = c(pca = 5, vae = 64),
  display = list(pca = c(NULL, NULL), vae = c(NULL, NULL)),
  platform = c("Visium", "ST"),
  diverging = FALSE,
  low = NULL,
  high = NULL,
  mid = NULL,
  color = NULL,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment as the output of \code{spatialPreprocess}.}

\item{datatype}{Type of data to plot, either features extracted from VAE,
the PCA features extracted from the results of VAE, or both.}

\item{res}{Features from which resolution to plot.}

\item{d}{A named vector for the numbers of features to plot.}

\item{display}{An optional list of vectors, each vector have two numbers
specifying the number of rows and columns the feature plots to be
organized.}

\item{platform}{Spatial sequencing platform. Used to determine spot layout
and neighborhood structure (Visium = hex, ST = square).}

\item{diverging}{If true, use a diverging color gradient in
\code{featurePlot()} (e.g. when plotting a fold change) instead of a
sequential gradient (e.g. when plotting expression).}

\item{low, mid, high}{Optional hex codes for low, mid, and high values of the
color gradient used for continuous spot values.}

\item{color}{Optional hex code to set color of borders around spots. Set to
\code{NA} to remove borders.}

\item{...}{Additional arguments for \code{geom_polygon()}. \code{size}, to
specify the linewidth of these borders, is likely the most useful.}
}
\value{
Returns a list of ggplot objects, named either \code{spot},
  \code{subspot}, or both.
}
\description{
Plot features extracted from images.
}
\seealso{
Other spatial plotting functions: 
\code{\link{clusterPlot}()},
\code{\link{featurePlot}()}
}
\concept{spatial plotting functions}
