% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{init.cluster}
\alias{init.cluster}
\title{Initialize cluster}
\usage{
init.cluster(Y, q, init = NULL, init.method = c("mclust", "kmeans"))
}
\arguments{
\item{Y}{PCs of SingleCellExperiment}

\item{q}{The number of clusters}

\item{init}{Vector of initial cluster assignments}

\item{init.method}{Initialization clustering algorithm}
}
\value{
Vector of initial cluster assignments
}
\description{
Initialize cluster
}
\keyword{internal}
