% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{lm.feature}
\alias{lm.feature}
\title{Predict feature vectors from enhanced PCs by linear model}
\usage{
lm.feature(X, X.enhanced, Y, feature_names)
}
\arguments{
\item{X}{Original PCs}

\item{X.enhanced}{Enhanced PCs}

\item{Y}{Original log expression.}

\item{feature_names}{List of genes/features to predict expression/values for.}
}
\value{
Vector of enhanced features.
}
\description{
Predict feature vectors from enhanced PCs by linear model
}
\keyword{internal}
