% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_spot_vertices}
\alias{make_spot_vertices}
\title{Make vertices outlining spots/subspots for geom_polygon()}
\usage{
make_spot_vertices(spot_positions, vertex)
}
\arguments{
\item{spot_positions}{Dataframe with row and col of each spots or subspots}

\item{vertex}{vertices of each spots or subspots}
}
\value{
Data frame with vertices outlining spots/subspot
}
\description{
Make vertices outlining spots/subspots for geom_polygon()
}
