% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_cluster.R
\name{make_subplots}
\alias{make_subplots}
\title{Add subspot labels and offset row/col locations.}
\usage{
make_subplots(sce, platform = c("Visium", "ST"))
}
\arguments{
\item{sce}{Original sce}

\item{platform}{If "Visium", make 6 suplots; if "ST", make  subplots.}
}
\value{
Data frame with added subplot names, parent spot indices, subplot indice,
and offset row/column coordinates.
}
\description{
Subspots are stored as (1.1, 2.1, 3.1, ..., 1.2, 2.2, 3.2, ...)
}
\keyword{internal}
