% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Find the mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
mode Numeric scalar, most frequent element in x
}
\description{
Used for finding the most frequent cluster for each z
}
\keyword{internal}
