% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlinePlot.R
\name{outlinePlot}
\alias{outlinePlot}
\title{Add outlines to a cluster/feature plot
TODO add `outline` and `outline_groups` (with default all) to c/fPlot}
\usage{
outlinePlot(
  sce,
  base_plot,
  outline_group = "spatial.cluster",
  outline_values = NULL,
  outline_title = "Cluster",
  outline_size = 1.2,
  outline_nudge = 0.5
)
}
\arguments{
\item{sce}{SingleCellExperiment.}

\item{base_plot}{Base cluster/feature plot to add outlines to.}

\item{outline_group}{Group to outline; i.e. a categorical column in
\code{colData(sce)}.}

\item{outline_values}{Subset of values in \code{outline_group} to outline; by
default all groups/clusters will be outlined.}

\item{outline_title}{Title of outline groups in plot legend.}

\item{outline_size}{Width of outline.}

\item{outline_nudge}{Scale the distance each vertex in the outline is nudged
towards the interior of the cluster.}
}
\value{
ggplot object
}
\description{
Add outlines to a cluster/feature plot
TODO add `outline` and `outline_groups` (with default all) to c/fPlot
}
