% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess SingleCellExperiemnt for BayesSpace}
\usage{
preprocess(
  sce,
  platform = c("Visium", "ST"),
  n.PCs = 15,
  n.HVGs = 2000,
  skip.PCA = FALSE,
  log.normalize = TRUE,
  assay.log = "logcounts"
)
}
\arguments{
\item{sce}{SingleCellExperiment to preprocess}

\item{platform}{Spatial sequencing platform.}

\item{n.PCs}{Number of principal components to compute.}

\item{n.HVGs}{Number of highly variable genes to run PCA upon.}

\item{skip.PCA}{Whether to skip PCA if dimentionality reduction was previously computed.}

\item{log.normalize}{Whether to log-normalize the input data with scater.}

\item{assay.log}{Name of assay in \code{sce} containing normalized counts.}
}
\value{
SingleCellExpriment with PCA and BayesSpace metadata
}
\description{
Add metadata for further analysis, and optionally performs PCA on log-normalized expression on top HVGs.
}
