% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readChain.R
\name{readChain}
\alias{readChain}
\title{Read MCMC chain associated with a BayesSpace clustering or enhancement}
\usage{
readChain(sce, params = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment with a file path stored in its metadata}

\item{params}{List of model parameters to read}
}
\description{
BayesSpace stores the MCMC chain associated with a clustering or enhancement 
on disk in an HDF5 file. The `readChain` function reads any parameters 
specified by the user into a `coda::mcmc` object compatible with TidyBayes.
}
\details{
To interact with the HDF5 file directly, obtain the filename from the 
SingleCellExperiment's metadata: `metadata(sce)$chain.h5`. Each parameter is 
stored as a separate dataset in the file, and is represented as a matrix of 
size (n_iterations x n_parameter_indices).
}
