% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVisium.R
\name{readVisium}
\alias{readVisium}
\title{Load a Visium spatial dataset as a SingleCellExperiment}
\usage{
readVisium(dirname)
}
\arguments{
\item{dirname}{Path to spaceranger output directory (e.g. "sampleID/outs/").
This directory must contain the counts matrix and feature/barcode TSVs in
\code{filtered_feature_bc_matrix/}, and the spot positions at
\code{spatial/tissue_positions_list.csv}. (These are default locations for
spaceranger outputs.)}
}
\value{
SingleCellExperiment containing the counts matrix in \code{counts}
  and spatial data in \code{colData}. Array coordinates for each spot are
  stored in columns \code{row} and \code{col}, while image coordinates are
  stored in columns \code{imagerow} and \code{imagecol}.
}
\description{
Load a Visium spatial dataset as a SingleCellExperiment
}
\details{
We store two variables associated with downstream BayesSpace
  functions in the SingleCellExperiment's metadata.
  \code{BayesSpace.platform} is set to "Visium", and is used to determine
  spot layout and neighborhood structure, and \code{BayesSpace.is_enhanced}
  is set to \code{FALSE} to denote the object contains spot-level data.
}
