% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runGenesetPCA}
\alias{runGenesetPCA}
\title{Run 'blocked' PCA on HVGs and gene sets of interest}
\usage{
runGenesetPCA(expr, n_pcs = 20, n_hvgs = 2000, genesets = list())
}
\arguments{
\item{expr}{Expression/counts (matrix or SingleCellExperiment)}

\item{n_pcs}{Number of PCs per gene set (TODO: optionally specify list, n 
per gene set)}

\item{n_hvgs}{Number of HVGs to use in HVG set}

\item{genesets}{List of other gene sets to include}
}
\value{
Returns list with names:
  * \code{x} Concatenated principal components from each gene set
  * \code{rotation} Rotation matrix (block diagonal for each set of PCs)
  * \code{genesets} List of gene sets in each block
}
\description{
PCA is computed for each specified gene set. The PCs from each gene set are 
concatenated into a single matrix, and their rotation matrices are combined
into a block diagonal rotation matrix.
}
\details{
If a set of highly variable genes (HVGs) is not included in the list of 
gene sets (using the name \code{hvgs}), the top \code{n_hvgs} will be
computed using scran and added to the list of gene sets.
}
