% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSpotlight.R
\name{runSPOTlight}
\alias{runSPOTlight}
\alias{runSPOTlight,ANY,ANY-method}
\alias{runSPOTlight,SummarizedExperiment,ANY-method}
\alias{runSPOTlight,SummarizedExperiment,SummarizedExperiment-method}
\title{Run SPOTlight cell type deconvolution on a spatial SCE}
\usage{
runSPOTlight(spatial, scRNA, ...)

\S4method{runSPOTlight}{ANY,ANY}(
  spatial,
  scRNA,
  sctransform.spatial = TRUE,
  sctransform.scRNA = FALSE,
  ...
)

\S4method{runSPOTlight}{SummarizedExperiment,ANY}(spatial, scRNA, ..., spatial.assay = "counts")

\S4method{runSPOTlight}{SummarizedExperiment,SummarizedExperiment}(
  spatial,
  scRNA,
  ...,
  spatial.assay = "counts",
  scRNA.assay = "counts"
)
}
\arguments{
\item{spatial}{SingleCellExperiment or expression matrix from the spatial
transcriptomic experiment.}

\item{scRNA}{SingleCellExperiment or expression matrix from a reference
single-cell RNA-seq experiment.}

\item{sctransform.spatial}{Run sctransform to normalize spatial expression.
Recommended by SPOTlight developers, but set to false if providing
log-normalized counts. (Will reduce runtime.)}

\item{sctransform.scRNA}{Run sctransform to normalize scRNA expression.
Recommended by SPOTlight developers, but set to false if providing
normalized counts. (Published reference datasets may only provide
normalized counts.)}

\item{cell_types}{Name of column in \code{colData(scRNA)} containing cell
type labels, or a list of labels.}
}
\description{
desc
}
