% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{SpatialCluster}
\alias{SpatialCluster}
\title{Spatial clustering}
\usage{
SpatialCluster(
  sce,
  q,
  d,
  platform = c("Visium", "ST"),
  init = NULL,
  init.method = c("mclust", "kmeans"),
  mu0 = NULL,
  Lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  gamma = NULL,
  ncores = 2,
  n.rep = 1000,
  save_para = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the spatial data.}

\item{q}{The numbers of clusters.}

\item{d}{The Number of top principal components to use when clustering.}

\item{platform}{Spatial transcriptomic platform.}

\item{init}{Vector of initial cluster assignments}

\item{init.method}{If \code{init} is not provided, cluster the top \code{d}
PCs with this method to obtain initial cluster assignments.}

\item{mu0}{Prior mean hyperparameter for mu. If not provided, mu0 is set to
the mean of PCs over all spots.}

\item{Lambda0}{Prior precision hyperparam for mu. If not provided, lambda0
is set to a diagonal matrix \eqn{0.01 I}.}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}

\item{gamma}{Smoothing parameter. Defaults to 2 for \code{platform="ST"} and
3 for \code{platform="Visium"}.}

\item{ncores}{The number of cores used for parallel calculation.}

\item{n.rep}{The number of MCMC iterations.}

\item{save_para}{If true, save the parameter mu, w and Lambda apart from z.}
}
\value{
Returns a modified \code{sce} with cluster assignments
stored in \code{colData} under the name \code{spatial.cluster},
and (optional) the list of clustering parameters,
stored BayesSpace metadata under the name \code{para}.
}
\description{
Cluster a spatial expression dataset.
}
