% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{spatialEnhance}
\alias{spatialEnhance}
\title{Enhance spot resolution}
\usage{
spatialEnhance(
  sce,
  q,
  platform = c("Visium", "ST"),
  use.dimred = "PCA",
  d = 15,
  init = NULL,
  init.method = c("spatialCluster", "mclust", "kmeans"),
  model = c("t", "normal"),
  nrep = 2e+05,
  gamma = 3,
  mu0 = NULL,
  lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  save.chain = FALSE,
  chain.fname = NULL,
  jitter_scale = 5,
  jitter_prior = 0.3,
  verbose = FALSE
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the spatial data.}

\item{q}{The number of clusters.}

\item{platform}{Spatial transcriptomic platform. Specify 'Visium' for hex 
lattice geometry or 'ST' for square lattice geometry.}

\item{use.dimred}{Name of a reduced dimensionality result in 
\code{reducedDims(sce)}. If provided, cluster on these features directly.}

\item{d}{Number of top principal components to use when clustering.}

\item{init}{Initial cluster assignments for spots.}

\item{init.method}{If \code{init} is not provided, cluster the top \code{d} 
PCs with this method to obtain initial cluster assignments.}

\item{model}{Error model. ('normal' or 't')}

\item{nrep}{The number of MCMC iterations.}

\item{gamma}{Smoothing parameter. (Values in range of 1-3 seem to work well.)}

\item{mu0}{Prior mean hyperparameter for mu. If not provided, mu0 is set to
the mean of PCs over all spots.}

\item{lambda0}{Prior precision hyperparam for mu. If not provided, lambda0
is set to a diagonal matrix \eqn{0.01 I}.}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}

\item{save.chain}{If true, save the MCMC chain to an HDF5 file.}

\item{chain.fname}{File path for saved chain. Tempfile used if not provided.}

\item{jitter_scale}{Controls the amount of jittering. Small amounts of 
jittering are more likely to be accepted but result in exploring the space
more slowly. We suggest tuning \code{jitter_scale} so that Ychange is on 
average around 30\%.}

\item{jitter_prior}{(TODO: re-define) Prior precision (1/variance) of the
actual spot-level expression value in PC space. We suggest making c larger
if the jittered values are not expected to vary much from the overall mean
of the spot.}

\item{verbose}{Log progress to stderr.}
}
\value{
Returns a new SingleCellExperiment object. By default, the 
  \code{assays} of this object are empty, and the enhanced resolution PCs 
  are stored as a reduced dimensionality result accessible with
  \code{reducedDim(sce, 'PCA')}.
}
\description{
Backend calls iterate_deconv(), written in Rcpp.
Inputs are the same as \code{spatialCluster()} except you have to specify
xdist and ydist instead of total dist...(maybe would be better to change
\code{spatialCluster} to match this)
}
\details{
TODO: add details on subspot coordinates and linking to parent spots
}
\examples{
set.seed(149)
sce <- exampleSCE()
sce <- spatialCluster(sce, 7, nrep=200)
enhanced <- spatialEnhance(sce, 7, init=sce$spatial.cluster, nrep=200)

}
