% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_cluster.R
\name{subplot.offsets}
\alias{subplot.offsets}
\title{Define offsets for each subspot layout.}
\usage{
subplot.offsets(platform = c("Visium", "ST"))
}
\arguments{
\item{platform}{If "Visium", devide hex spots into 6 triangular subspots ; if
"ST", devide square spots into 9 squares.}
}
\value{
Matrix of x and y offsets, one row per subspot
}
\description{
Hex spots are divided into 6 triangular subspots, square spots are divided
into 9 squares. Offsets are relative to the spot center.
}
\keyword{internal}
