% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tri_spots_vertices}
\alias{tri_spots_vertices}
\title{Make vertices for each triangle subspot of a hex}
\usage{
tri_spots_vertices(sce, x = "sub.col", y = "sub.row", fill = "spatial.cluster")
}
\value{
Table of (x.pos, y.pos, spot, fill); where \code{spot} groups the
  vertices outlining the spot's border
}
\description{
Make vertices for each triangle subspot of a hex
}
\keyword{internal}
