% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{update.mu}
\alias{update.mu}
\title{Update parameter mu}
\usage{
\method{update}{mu}(j, z, Y, mu0, Lambda0, Lambda, w, mu)
}
\arguments{
\item{j}{Update mean of cluster j}

\item{z}{Vector of cluster assignments}

\item{Y}{PCs of SingleCellExperiment}

\item{mu0}{Prior mean hyperparameter for mu. If not provided, mu0 is set to
the mean of PCs over all spots.}

\item{Lambda0}{Prior precision hyperparam for mu. If not provided, lambda0
is set to a diagonal matrix \eqn{0.01 I}.}

\item{Lambda}{Precision matrix}

\item{w}{Proportion factor}

\item{mu}{Mean of cluster}
}
\value{
New mean of cluster j
}
\description{
Update parameter mu
}
\keyword{internal}
