% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{update.w}
\alias{update.w}
\title{Update parameter w}
\usage{
\method{update}{w}(j, z, Y, d, v, Lambda, mu, w)
}
\arguments{
\item{j}{Update jth proportion factor}

\item{z}{Vector of cluster assignments}

\item{Y}{Vector of cluster assignments}

\item{d}{Dimention of PCs}

\item{v}{A fixed degree of freedom}

\item{Lambda}{Precision matrix}

\item{mu}{Mean of cluster}

\item{w}{Proportion factor}
}
\value{
New jth proportion factor
}
\description{
Update parameter w
}
\keyword{internal}
