% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster.R
\name{update.z}
\alias{update.z}
\title{Update parameter z}
\usage{
\method{update}{z}(j, z, Y, q, spots.neighbors, gamma, Lambda, mu, w)
}
\arguments{
\item{j}{Update jth cluster assignments}

\item{z}{Vector of cluster assignments}

\item{Y}{Vector of cluster assignments}

\item{q}{Number of clusters}

\item{spots.neighbors}{List of neighbor indices for each spot}

\item{gamma}{Smoothing parameter. Defaults to 2 for \code{platform="ST"} and
3 for \code{platform="Visium"}.}

\item{Lambda}{Precision matrix}

\item{mu}{Mean of cluster}

\item{w}{Proportion factor}
}
\value{
New jth cluster assignments
}
\description{
Update parameter z
}
\keyword{internal}
