% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{xgboost.feature}
\alias{xgboost.feature}
\title{Predict feature vectors from enhanced PCs by xgboost}
\usage{
xgboost.feature(X, X.enhanced, Y, feature_names, nrounds, train.n)
}
\arguments{
\item{X}{Original PCs}

\item{X.enhanced}{Enhanced PCs}

\item{Y}{Original log expression.}

\item{feature_names}{List of genes/features to predict expression/values for.}

\item{nrounds}{Nonnegative integer to set the \code{nrounds} parameter for xgboost.
If \code{nrounds} is set to 0, the parameter will be tuned using a train-test split.
We recommend tuning \code{nrounds} for improved feature prediction, but note this will increase
runtime.}

\item{train.n}{Number of spots to use in the training dataset for tuning
nrounds. By default, 2/3 the total number of spots are used.}
}
\value{
Vector of enhanced features.
}
\description{
Predict feature vectors from enhanced PCs by xgboost
}
\keyword{internal}
