% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addPCA}
\alias{addPCA}
\title{Add PCA output to a SingleCellExperiment}
\usage{
addPCA(sce, assay.type, pca.method, d = 15)
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{assay.type}{Assay in \code{sce} to compute PCA on}

\item{pca.method}{PCA method to apply}

\item{d}{Number of principal components to keep}
}
\value{
Returns \code{sce} with PCs added to its \code{reducedDims}
}
\description{
Supports 'basic' PCA via scater, denoised PCA via scran, and blocked geneset
PCA. Results will be stored in \code{reducedDim(sce, 'PCA')}.
}
\keyword{internal}
