% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpace.R
\docType{package}
\name{BayesSpace}
\alias{BayesSpace}
\title{BayesSpace: A package for processing spatial transcriptomes}
\description{
The BayesSpace package provides two primary functions: `cluster()` and `deconvolve()`.
}
\section{Clustering}{

The clustering algorithm groups neighboring spots together by expression profile.
}

\section{Deconvolution}{

The deconvolution algorithm deconvolves spot-level expression signals into imputed cell-level signals.
}

