% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpace.R
\docType{package}
\name{BayesSpace}
\alias{BayesSpace}
\title{BayesSpace: A package for processing spatial transcriptomes}
\description{
TODO: flesh out package description (or link to vignette a la scran)
}
\details{
The BayesSpace package provides two primary functions:
\code{spatialCluster()} and \code{spatialEnhance()}.
}
\section{Clustering}{

The clustering algorithm groups neighboring spots together by expression
profile.
}

\section{Enhancement}{

The enhancement algorithm deconvolves spot-level expression into imputed
cell-level signals.
}

