% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{cluster}
\alias{cluster}
\title{Wrapper around C++ \code{iterate_*()} functions}
\usage{
cluster(
  Y,
  positions,
  radius,
  q,
  init = rep(1, nrow(Y)),
  model = c("t", "normal"),
  precision = c("equal", "variable"),
  mu0 = colMeans(Y),
  lambda0 = diag(0.01, nrow = ncol(Y)),
  gamma = 3,
  alpha = 1,
  beta = 0.01,
  nrep = 1000
)
}
\value{
List of clustering parameter values at each iteration
}
\description{
Wrapper around C++ \code{iterate_*()} functions
}
\keyword{internal}
