% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution of spots into cells}
\usage{
deconvolve(
  Y,
  positions,
  nrep = 1000,
  gamma = 2,
  xdist,
  ydist,
  q,
  init,
  model = "normal",
  platform = c("Visium", "ST"),
  verbose = TRUE,
  jitter_scale = 5,
  c = 0.01,
  mu0 = colMeans(Y),
  lambda0 = diag(0.01, nrow = ncol(Y)),
  alpha = 1,
  beta = 0.01
)
}
\arguments{
\item{Y}{A matrix or dataframe with 1 row per spot and 1 column per outcome 
(e.g. principal component).}

\item{positions}{A matrix or dataframe with two columns (x, y) that gives 
the spatial coordinates of the spot.}

\item{nrep}{The maximum number of mcmc iterations.}

\item{gamma}{Smoothing parameter. Values in range of 1-3 generally work well.}

\item{xdist}{The distance along x-axis between neighboring spots.}

\item{ydist}{The distance along y-axis between neighboring spots.}

\item{q}{The number of clusters.}

\item{init}{Initial cluster assignments (z's). Must be a vector of length 
equal to the number of rows of Y and positions.}

\item{model}{Error model. ("normal" or "t")}

\item{platform}{Spatial transcriptomic platform. Specify "Visium" for hex 
lattice geometry or "ST" for square lattice geometry.}

\item{verbose}{Log progress to stderr.}

\item{jitter_scale}{Controls the amount of jittering. Small amounts of 
jittering are more likely to be accepted but result in exploring the space
more slowly. We suggest tuning `jitter_scale` so that Ychange is on 
average around 30\%.}

\item{c}{Prior precision (1/variance) of the actual spot-level expression 
value in PC space. We suggest making c larger if the jittered values are
not expected to vary much from the overall mean of the spot.}

\item{mu0}{Prior mean hyperparameter for mu.}

\item{lambda0}{Prior precision hyperparam for mu.}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}
}
\value{
out TODO: specify
}
\description{
Backend calls iterate_deconv(), written in Rcpp.
Inputs are the same as `cluster()` except you have to specify xdist and ydist 
instead of total dist...(maybe would be better to change `cluster()` to match this)
}
