% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution of spots into cells}
\usage{
deconvolve(
  Y,
  positions,
  nrep = 1000,
  every = 1,
  gamma = 2,
  xdist,
  ydist,
  q,
  init,
  seed = 100,
  mu0 = colMeans(Y),
  lambda0 = diag(0.01, nrow = ncol(Y)),
  alpha = 1,
  beta = 0.01
)
}
\arguments{
\item{Y}{A matrix or dataframe with 1 row per spot and 1 column per outcome (e.g. principal component)}

\item{positions}{A matrix or dataframe with two columns (x, y) that gives the spatial coordinates of the spot}

\item{nrep}{The maximum number of mcmc iterations}

\item{every}{TODO: define}

\item{gamma}{Smoothing parameter. Values in range of 1-3 seem to work well generally}

\item{xdist}{The distance along x-axis between neighboring spots}

\item{ydist}{The distance along y-axis between neighboring spots}

\item{q}{The number of clusters}

\item{init}{Initial cluster assignments (z's). Must be a vector of length equal to the number of rows of Y and positions}

\item{seed}{Random state seed}

\item{mu0}{Prior mean hyperparameter for mu}

\item{lambda0}{Prior precision hyperparam for mu}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda}

\item{beta}{Hyperparameter for Wishart distributed precision lambda}
}
\description{
Backend calls iterate_deconv(), written in Rcpp.
Inputs are the same as `cluster()` except you have to specify xdist and ydist 
instead of total dist...(maybe would be better to change `cluster()` to match this)
}
