% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{deconvolve}
\alias{deconvolve}
\title{Wrapper around C++ \code{iterate_deconv()} function}
\usage{
deconvolve(
  Y,
  positions,
  nrep = 1000,
  gamma = 2,
  xdist,
  ydist,
  q,
  init,
  model = "normal",
  platform = c("Visium", "ST"),
  verbose = TRUE,
  jitter_scale = 5,
  jitter_prior = 0.01,
  mu0 = colMeans(Y),
  lambda0 = diag(0.01, nrow = ncol(Y)),
  alpha = 1,
  beta = 0.01
)
}
\description{
Wrapper around C++ \code{iterate_deconv()} function
}
\keyword{internal}
