% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{.init_cluster}
\alias{.init_cluster}
\title{Initialize cluster assignments}
\usage{
.init_cluster(Y, q, init = NULL, init.method = c("mclust", "kmeans"))
}
\arguments{
\item{Y}{Representation of reduced dimensions}

\item{q}{Number of clusters}

\item{init}{Vector of initial cluster assignments}

\item{init.method}{Initialization clustering algorithm}
}
\value{
Vector of cluster assignments.
}
\description{
Initialize cluster assignments
}
\keyword{internal}
