% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{.init_cluster}
\alias{.init_cluster}
\title{Initialize cluster assignments}
\usage{
.init_cluster(sce, q, inputs, init = NULL, init.method = c("mclust", "kmeans"))
}
\arguments{
\item{sce}{SingleCellExperiment}

\item{q}{Number of clusters}

\item{inputs}{Results from .prepare_inputs() (TODO: store this in sce)}

\item{init}{Vector of initial cluster assignments}

\item{init.method}{Initialization clustering algorithm}
}
\value{
Modified sce with initial cluster assignments stored in
  colData$cluster.init (TODO: return vector instead)
}
\description{
Initialize cluster assignments
}
\keyword{internal}
