% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEnhance.R
\name{.make_subspot_coldata}
\alias{.make_subspot_coldata}
\title{Add subspot labels and offset row/col locations before making enhanced SCE.}
\usage{
.make_subspot_coldata(
  cdata,
  sce,
  subspot_neighbors,
  platform,
  nsubspots.per.edge = 3
)
}
\arguments{
\item{cdata}{Table of colData (imagerow and imagecol; from deconv$positions)}

\item{sce}{Original sce (to obtain number of spots and original row/col)}

\item{subspot_neighbors}{Neighbors for subspots}

\item{platform}{Spatial transcriptomic platform}

\item{nsubspots.per.edge}{Number of subspots per edge if the spot is squared}
}
\value{
Data frame with added subspot names, parent spot indices, and offset
  row/column coordinates
}
\description{
Subspots are stored as (1.1, 2.1, 3.1, ..., 1.2, 2.2, 3.2, ...)
}
\keyword{internal}
