% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.make_triangle_subspots}
\alias{.make_triangle_subspots}
\title{Make vertices for each triangle subspot of a hex}
\usage{
.make_triangle_subspots(cdata, fill.col = "spatial.cluster")
}
\description{
Make vertices for each triangle subspot of a hex
}
\keyword{internal}
