% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prepare_inputs}
\alias{.prepare_inputs}
\title{Prepare cluster/deconvolve inputs from SingleCellExperiment object}
\usage{
.prepare_inputs(
  sce,
  use.dimred = "PCA",
  d = 15,
  positions = NULL,
  position.cols = c("imagecol", "imagerow"),
  radius = NULL,
  xdist = NULL,
  ydist = NULL
)
}
\value{
List of PCs, names of columns with x/y positions, and inter-spot
  distances
}
\description{
Prepare cluster/deconvolve inputs from SingleCellExperiment object
}
\keyword{internal}
