% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{.select_spot_positions}
\alias{.select_spot_positions}
\title{Helper to extract x, y, fill ID from colData}
\usage{
.select_spot_positions(
  cdata,
  x = "array_col",
  y = "array_row",
  fill = "spatial.cluster"
)
}
\value{
Dataframe of (x.pos, y.pos, fill) for each spot
}
\description{
Helper to extract x, y, fill ID from colData
}
\keyword{internal}
