% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceFeatures.R
\name{enhanceFeatures}
\alias{enhanceFeatures}
\title{Predict log-normalized expression vectors from deconvolved PCs using linear 
regression.}
\usage{
enhanceFeatures(
  sce.enhanced,
  sce.ref,
  use.dimred = "PCA",
  assay.type = "logcounts",
  altExp.type = NULL,
  feature.matrix = NULL,
  ...
)
}
\arguments{
\item{sce.enhanced}{SingleCellExperiment object with enhanced PCs.}

\item{sce.ref}{SingleCellExperiment object with original PCs and expression..}

\item{use.dimred}{Name of dimension reduction to use.}

\item{assay.type}{Expression matrix in \code{assays(sce.ref)} to predict.}

\item{altExp.type}{Expression matrix in \code{altExps(sce.ref)} to predict.
Overrides \code{assay.type} if specified.}

\item{feature.matrix}{Expression/feature matrix to predict, if not directly
attached to \code{sce.ref}. Must have columns corresponding to the spots in
\code{sce.ref}. Overrides \code{assay.type} and \code{altExp.type} if
specified.}

\item{features}{List of genes/features to predict expression/values for.}

\item{model}{Model used to predict enhanced values.}
}
\value{
If \code{assay.type} or \code{altExp.type} are specified, the
  enhanced features are stored in the corresponding slot of
  \code{sce.enhanced} and the modified SingleCellExperiment object is
  returned.
  
  If \code{feature.matrix} is specified, the enhanced features are returned
  directly as a matrix.
}
\description{
Predict log-normalized expression vectors from deconvolved PCs using linear 
regression.
}
\details{
TODO add details
}
\examples{
set.seed(149)
sce <- exampleSCE()
sce <- spatialCluster(sce, 7)
enhanced <- spatialEnhance(sce, 7, init=sce$spatial.cluster)
enhanceFeatures(enhanced, sce, assay.type="logcounts")

}
