% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{EnhanceFeatures}
\alias{EnhanceFeatures}
\title{Predict feature vectors from enhanced PCs.}
\usage{
EnhanceFeatures(
  sce.enhanced,
  sce,
  feature_names,
  model = c("xgboost", "lm"),
  nrounds = 0,
  train.n = round(ncol(sce) * 2/3)
)
}
\arguments{
\item{sce.enhanced}{SingleCellExperiment object with enhanced PCs.}

\item{sce}{SingleCellExperiment object with original PCs and log expression.}

\item{feature_names}{List of genes/features to predict expression/values for.}

\item{model}{Model used to predict enhanced values.}

\item{nrounds}{Nonnegative integer to set the \code{nrounds} parameter for xgboost.
If \code{nrounds} is set to 0, the parameter will be tuned using a train-test split.
We recommend tuning \code{nrounds} for improved feature prediction, but note this will increase
runtime.}

\item{train.n}{Number of spots to use in the training dataset for tuning
nrounds. By default, 2/3 the total number of spots are used.}
}
\value{
Th enhanced features are stored in the \code{logcounts} of
  \code{sce.enhanced} and the modified SingleCellExperiment object is
  returned.
}
\description{
Predict feature vectors from enhanced PCs.
}
