% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleSCE}
\alias{exampleSCE}
\title{Create minimal \code{SingleCellExperiment} for documentation examples.}
\usage{
exampleSCE(n_PCs = 15, nrow = 10, ncol = 20)
}
\arguments{
\item{n_PCs}{Number of principal components to simulate}

\item{nrow}{Number of rows of spots}

\item{ncol}{Number of columns of spots}
}
\value{
A SingleCellExperiment object with no assays, a set of random PCs in
\code{reducedDim(sce, "PCA")}, and positional data in \code{colData}. Spots
are distributed over an (\code{nrow} x \code{ncol}) rectangle.
}
\description{
Create minimal \code{SingleCellExperiment} for documentation examples.
}
\details{
Inspired by scuttle's \code{mockSCE}.
}
\examples{
set.seed(149)
sce <- exampleSCE()

}
