% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRDS}
\alias{getRDS}
\title{Download a processed sample from our S3 bucket}
\usage{
getRDS(dataset, sample)
}
\arguments{
\item{dataset}{Dataset identifier (TODO: add function to list datasets/samples)}

\item{sample}{Sample identifier}
}
\value{
sce A SingleCellExperiment with positional information in colData and
  PCs based on the top 2000 HVGs
}
\description{
Download a processed sample from our S3 bucket
}
