% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictExpression.R
\name{predictExpression}
\alias{predictExpression}
\title{Predict log-normalized expression vectors from deconvolved PCs using linear regression}
\usage{
predictExpression(
  sce,
  newdata,
  dimred = "PCA",
  genes = rownames(sce),
  components = ncol(newdata)
)
}
\arguments{
\item{sce}{SingleCellExperiment object with original PCs and logcounts}

\item{newdata}{Deconvolved PCs (rows are subspots, columns are PCs)}

\item{dimred}{Name of dimension reduction to use}

\item{genes}{List of genes to predict expression for}

\item{components}{Number of reduced dimensions (PCs) to use}
}
\value{
Returns list with names:
        * `expression` - Deconvolved expression values. (Rows are genes, columns are subspots)
        * `r2` - Percent of variation in original gene expression explained by PCs
}
\description{
Backend calls iterate_deconv(), written in Rcpp.
Inputs are the same as `cluster()` except you have to specify xdist and ydist 
instead of total dist...(maybe would be better to change `cluster()` to match this)
}
