% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCluster.R
\name{spatialCluster}
\alias{spatialCluster}
\title{Spatial clustering}
\usage{
spatialCluster(
  sce,
  q,
  use.dimred = "PCA",
  d = 15,
  positions = NULL,
  position.cols = c("imagecol", "imagerow"),
  init = NULL,
  init.method = c("kmeans"),
  radius = NULL,
  model = c("normal", "t"),
  precision = c("equal", "variable"),
  nrep = 1000,
  gamma = 2,
  mu0 = NULL,
  lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  save.chain = FALSE,
  chain.fname = NULL
)
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the spatial data.}

\item{q}{The number of clusters.}

\item{use.dimred}{Name of a reduced dimensionality result in 
\code{reducedDims(sce)}. If provided, cluster on these features directly.}

\item{d}{Number of top principal components to use when clustering.}

\item{positions}{A matrix or dataframe with two columns (x, y) that gives
the spatial coordinates of each spot.}

\item{position.cols}{If \code{positions} is not provided, use these columns 
from \code{colData(sce)} as spatial coordinates (x, y).}

\item{init}{Initial cluster assignments for spots.}

\item{init.method}{If \code{init} is not provided, cluster the top \code{d} 
PCs with this method to obtain initial cluster assignments.}

\item{radius}{The maximum (L1) distance for two spots to be considered
neighbors. If not provided, the distance will be estimated using
\code{lm(image.coord ~ array.coord)}.}

\item{model}{Error model. ('normal' or 't')}

\item{precision}{Covariance structure. ('equal' or 'variable' for EEE and 
VVV covariance models, respectively.)}

\item{nrep}{The number of MCMC iterations.}

\item{gamma}{Smoothing parameter. (Values in range of 1-3 seem to work well.)}

\item{mu0}{Prior mean hyperparameter for mu. If not provided, mu0 is set to
the mean of PCs over all spots.}

\item{lambda0}{Prior precision hyperparam for mu. If not provided, lambda0
is set to a diagonal matrix \eqn{0.01 I}.}

\item{alpha}{Hyperparameter for Wishart distributed precision lambda.}

\item{beta}{Hyperparameter for Wishart distributed precision lambda.}

\item{save.chain}{If true, save the MCMC chain to an HDF5 file.}

\item{chain.fname}{File path for saved chain. Tempfile used if not provided.}
}
\value{
Returns a modified \code{sce} with cluster assignments stored in
  \code{colData} under the name \code{spatial.cluster}.
}
\description{
Cluster a spatial expression dataset.
}
\details{
TODO describe method in detail
  TODO add details or link to mcmcChain
}
