% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{spatialPlot}
\alias{spatialPlot}
\alias{clusterPlot}
\alias{enhancePlot}
\title{Plot cluster labels in spatial context}
\usage{
clusterPlot(sce, platform = c("Visium", "ST"), fill = "spatial.cluster")

enhancePlot(sce, platform = c("Visium", "ST"))
}
\arguments{
\item{sce}{SingleCellExperiment containing cluster assignments in
\code{sce$spatial.cluster}}

\item{platform}{Spatial sequencing platform. If "Visium", the hex spot layout
will be used, otherwise square spots will be plotted.}
}
\value{
Both functions return a \code{ggplot} object.
}
\description{
After running \code{spatialCluster()} or \code{spatialEnhance()}, we can
visualize the cluster assignments of each spot using \code{clusterPlot()} or
\code{enhancePlot()}, respectively.
}
\examples{
set.seed(149)
sce <- exampleSCE()
sce$spatial.cluster <- floor(runif(ncol(sce), 1, 4))
clusterPlot(sce)

}
