% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{spatialPlot}
\alias{spatialPlot}
\alias{clusterPlot}
\alias{enhancePlot}
\title{Plot cluster labels in spatial context}
\usage{
clusterPlot(sce, platform = c("Visium", "ST"))

enhancePlot(sce, platform = c("Visium", "ST"))
}
\arguments{
\item{sce}{A SingleCellExperiment object containing the spatial data and
cluster assignments in \code{sce$spatial.cluster}.}

\item{sce.enhanced, sce.ref}{The enhanced subspots and original spots,
respectively.}
}
\value{
Both functions return a ggplot object.
}
\description{
After running \code{spatialCluster()} or \code{spatialEnhance()}, we can
visualize the cluster assignments of each spot using \code{clusterPlot()} or
\code{enhancePlot()}, respectively.
}
\examples{
set.seed(149)
sce <- exampleSCE()
sce$spatial.cluster <- floor(runif(ncol(sce), 1, 4))
clusterPlot(sce)

}
